chore(config): Refactor build script and reorganize server file
- Updated the entry point in `build.mjs` from `"whServer.ts"` to `"src/whServer.ts"`.
- Renamed and moved `whServer.ts` to `src/whServer.ts`.
- Added license header, including a link to the TPS Technology website and copyright notice, in the `src/whServer.ts` file.
- Changed the import path of `mySecretFn` in `src/whServer.ts` from `"./secrets/mySecretFns"` to `"../secrets/mySecretFns"`.
- Updated the TODO comment in `src/whServer.ts` to suggest processing the webhook payload using `mySecretFn`.
- Adjusted the `include` path in `tsconfig.json` to reflect the new location of `whServer.ts`, now `src/whServer.ts`.

These modifications streamline the project structure by relocating the server file to a `src` directory, ensuring better organization. The update in `build.mjs` reflects this change. Additionally, the introduction of a license header in `whServer.ts` (now `src/whServer.ts`) clarifies licensing and ownership. The change in the TODO comment provides a clearer direction for future development.

Release-as: 1.0.0